Refactor `RdsDataSourceConfig` package structure and update dependencies for Spring Boot 4.x compatibility. Updated Maven configurations to align with the new version.
