+++ src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.jdbc.RdsDataSourceConfig
+org.springframework.boot.jdbc.autoconfigure.RdsDataSourceConfig
